convert all size_t variables to unsigned
This will ease serialization, reduce memory footprint, and increase
performance. In truthness using size_t was never really sensible for our
case, just muscle memory on my part

+++ src/game.c
@@ -41,7 +41,7 @@ void FreeInnerLogEntry(LogEntry entry) {
 
 GameLog DeepCopyGameLog(GameLog game_log) {
   LogEntry *log_entries = malloc(sizeof *game_log.items * game_log.count);
-  for (size_t i = 0; i < game_log.count; i++) {
+  for (unsigned i = 0; i < game_log.count; i++) {
     log_entries[i] = DeepCopyLogEntry(game_log.items[i]);
   }
 
@@ -56,7 +56,7 @@ GameLog DeepCopyGameLog(GameLog game_log) {
 }
 
 void FreeInnerGameLog(GameLog game_log) {
-  for (size_t i = 0; i < game_log.count; i++) {
+  for (unsigned i = 0; i < game_log.count; i++) {
     FreeInnerLogEntry(game_log.items[i]);
   }
   free(game_log.items);
@@ -107,7 +107,7 @@ BotsDA DeepCopyBotsDA(BotsDA bots) {
       .count = bots.count,
       .capacity = bots.count,
   };
-  for (size_t i = 0; i < bots.count; i++) {
+  for (unsigned i = 0; i < bots.count; i++) {
     new_bots.items[i] = DeepCopyBot(bots.items[i]);
   }
   return new_bots;
@@ -167,13 +167,13 @@ unsigned ParseMapFile(GameState *state, const char *map_path) {
   }
 
   char buf[256];
-  size_t file_line = 0;
+  unsigned file_line = 0;
   int bot_count = 0;
   state->bot_bit_set = 0;
 
   while (fgets(buf, sizeof buf, map_file)) {
     file_line += 1;
-    size_t len = strlen(buf);
+    unsigned len = strlen(buf);
     if (len == sizeof(buf) - 1 && buf[len - 1] != '\n') {
       nob_log(NOB_ERROR,
               "Map file contains lines longer then 256 characters and "
@@ -186,14 +186,14 @@ unsigned ParseMapFile(GameState *state, const char *map_path) {
 
     Planet planet;
     if (!ParsePlanetLine(buf, &planet)) {
-      nob_log(NOB_ERROR, "Invalid map file.\nSyntax error at line %zu.",
+      nob_log(NOB_ERROR, "Invalid map file.\nSyntax error at line %u.",
               file_line);
       exit(1);
     }
     if (planet.owner > MAX_BOT_AMOUNT) {
       nob_log(NOB_ERROR,
               "Map containes more owners then the max bot count. Encountered "
-              "in line: %zu\nOwner found: %d\nMax bot count: %d",
+              "in line: %u\nOwner found: %d\nMax bot count: %d",
               file_line, planet.owner, MAX_BOT_AMOUNT);
       exit(1);
     }
@@ -228,10 +228,10 @@ GameState MakeGame(const char *map_file_path, BotsDA bots, bool log) {
 
   // ----- MAP -----
   nob_log(NOB_INFO, "Loading map file from %s.", map_file_path);
-  size_t owner_count = ParseMapFile(&state, map_file_path);
+  unsigned owner_count = ParseMapFile(&state, map_file_path);
   if (owner_count != bots.count) {
     nob_log(NOB_ERROR,
-            "Provided map requires %zu player, yet %zu bots were given as "
+            "Provided map requires %u player, yet %u bots were given as "
             "arguments.",
             owner_count, bots.count);
     exit(1);
@@ -250,7 +250,7 @@ GameState MakeGame(const char *map_file_path, BotsDA bots, bool log) {
   return state;
 }
 
-void DisqualifyBot(GameState *state, size_t bot_idx) {
+void DisqualifyBot(GameState *state, unsigned bot_idx) {
   if (bot_idx >= state->bots.count) {
     nob_log(NOB_ERROR, "Attempted to disqualify non existent bot");
     exit(1);
@@ -265,12 +265,12 @@ void DisqualifyBot(GameState *state, size_t bot_idx) {
 
   state->remaining_bots--;
   nob_da_foreach(Planet, planet, &state->planets) {
-    if ((size_t)planet->owner == bot_idx + 1) {
+    if ((unsigned)planet->owner == bot_idx + 1) {
       planet->owner = 0;
     }
   }
   nob_da_foreach(Fleet, fleet, &state->fleets) {
-    if ((size_t)fleet->owner == bot_idx + 1) {
+    if ((unsigned)fleet->owner == bot_idx + 1) {
       *fleet = state->fleets.items[--state->fleets.count];
       fleet--;
     }
@@ -278,7 +278,7 @@ void DisqualifyBot(GameState *state, size_t bot_idx) {
 
   UnsetBit(state->bot_bit_set, bot_idx);
 
-  nob_log(NOB_INFO, "Disqualified bot %zu.", bot_idx);
+  nob_log(NOB_INFO, "Disqualified bot %u.", bot_idx);
 }
 
 static inline void PrintPlanet(FILE *file, Planet planet) {
@@ -296,19 +296,19 @@ static inline void PrintFleet(FILE *file, Fleet fleet) {
   fwrite(buf, sizeof *buf, len, file);
 }
 
-void sendMapToBot(GameState *state, size_t bot_idx) {
+void sendMapToBot(GameState *state, unsigned bot_idx) {
   if (bot_idx >= state->bots.count) {
     nob_log(NOB_ERROR, "ERROR: Attempting access to non-existent bot process");
     exit(1);
   }
   if (!subprocess_alive(state->bots.items[bot_idx].process)) {
-    nob_log(NOB_INFO, "Bot %zu has crashed.", bot_idx);
+    nob_log(NOB_INFO, "Bot %u has crashed.", bot_idx);
     DisqualifyBot(state, bot_idx);
     return;
   }
   FILE *bot_stdin = subprocess_stdin(state->bots.items[bot_idx].process);
 
-  WriteToLogFile("engine > player%zu: ", bot_idx + 1);
+  WriteToLogFile("engine > player%u: ", bot_idx + 1);
 
 #define MoveOwner(Type, entity)                                                \
   Type moved_##entity = *entity;                                               \
@@ -342,19 +342,19 @@ void sendMapToBot(GameState *state, size_t bot_idx) {
 
 // Return true if everythin went okay. Return false in case bot should be
 // disqualified.
-bool GetBotMessage(GameState *state, Nob_String_Builder *sb, size_t bot_idx) {
+bool GetBotMessage(GameState *state, Nob_String_Builder *sb, unsigned bot_idx) {
   if (bot_idx >= state->bots.count) {
     nob_log(NOB_ERROR, "Tried accessing a bot OOB.");
     exit(1);
   }
   if (!subprocess_alive(state->bots.items[bot_idx].process)) {
-    nob_log(NOB_INFO, "Bot %zu disqualified since it's process crashed.",
+    nob_log(NOB_INFO, "Bot %u disqualified since it's process crashed.",
             bot_idx);
     sb->count = 0;
     return false;
   }
 
-  const size_t max_chunk_length = 512;
+  const unsigned max_chunk_length = 512;
   sb->count = 0;
   bool message_ended = false;
 
@@ -371,7 +371,7 @@ bool GetBotMessage(GameState *state, Nob_String_Builder *sb, size_t bot_idx) {
                                sb->items + sb->count, sb->capacity - sb->count);
     if (received == 0) {
       if (nob_nanos_since_unspecified_epoch() - start > MAX_BOT_RESPONSE_TIME) {
-        nob_log(NOB_INFO, "Bot %zu disqualified for taking too long to reply.",
+        nob_log(NOB_INFO, "Bot %u disqualified for taking too long to reply.",
                 bot_idx);
         sb->count = 0;
         return false;
@@ -381,25 +381,25 @@ bool GetBotMessage(GameState *state, Nob_String_Builder *sb, size_t bot_idx) {
     }
     sb->count += received;
 
-    nob_log(NOB_DEBUG, "bot %zu sent: |%.*s|", bot_idx, (unsigned)sb->count,
+    nob_log(NOB_DEBUG, "bot %u sent: |%.*s|", bot_idx, (unsigned)sb->count,
             sb->items);
 
     // Excluding null terminator
-    const size_t delimeter_length = NOB_ARRAY_LEN(MESSAGE_DELIMETER) - 1;
+    const unsigned delimeter_length = NOB_ARRAY_LEN(MESSAGE_DELIMETER) - 1;
     // We need to check sb.count is at least `delimeter_length` to make sure
     // memcmp does not access OOB memory
     if (sb->count >= delimeter_length &&
         memcmp(sb->items + sb->count - delimeter_length, MESSAGE_DELIMETER,
                delimeter_length) == 0) {
       message_ended = true;
-      nob_log(NOB_DEBUG, "bot %zu message ended", bot_idx);
+      nob_log(NOB_DEBUG, "bot %u message ended", bot_idx);
     }
   }
 
   Nob_String_View sv = {sb->count, sb->items};
   while (sv.count > 0) {
     Nob_String_View line = nob_sv_chop_by_delim(&sv, '\n');
-    WriteToLogFile("player%zu > engine: %.*s\n", bot_idx + 1, (int)line.count,
+    WriteToLogFile("player%u > engine: %.*s\n", bot_idx + 1, (int)line.count,
                    line.data);
   }
   return true;
@@ -408,7 +408,7 @@ bool GetBotMessage(GameState *state, Nob_String_Builder *sb, size_t bot_idx) {
 // Return true if everythin went okay. Return false in case bot should be
 // disqualified.
 bool ParseBotFleets(GameState *state, Nob_String_View bot_message,
-                    size_t bot_idx) {
+                    unsigned bot_idx) {
   if (bot_message.count < 2 ||
       (bot_message.data[0] == 'g' && bot_message.data[1] == 'o')) {
     return true;
@@ -416,7 +416,7 @@ bool ParseBotFleets(GameState *state, Nob_String_View bot_message,
 
   while (bot_message.count > 1 && bot_message.data[0] != 'g' &&
          bot_message.data[1] != 'o') {
-    nob_log(NOB_DEBUG, "parsing bot %zu fleets", bot_idx);
+    nob_log(NOB_DEBUG, "parsing bot %u fleets", bot_idx);
     Fleet fleet;
     fleet.owner = bot_idx + 1;
     const char *start = bot_message.data;
@@ -440,7 +440,7 @@ bool ParseBotFleets(GameState *state, Nob_String_View bot_message,
     bot_message.count -= bot_message.data - start;
     bot_message = nob_sv_trim_left(bot_message);
 
-    if ((size_t)fleet.src_id >= state->planets.count) {
+    if ((unsigned)fleet.src_id >= state->planets.count) {
       nob_log(NOB_INFO, "Bot tried sending fleet from nonexistent planet.");
       return false;
     }
@@ -457,7 +457,7 @@ bool ParseBotFleets(GameState *state, Nob_String_View bot_message,
       nob_log(NOB_INFO,
               "Bot tried sending fleet from a planet it does not own.");
       return false;
-    } else if ((size_t)fleet.dst_id > state->planets.count) {
+    } else if ((unsigned)fleet.dst_id > state->planets.count) {
       nob_log(NOB_INFO, "Bot tried sending fleet to nonexistent planet.");
       return false;
     } else if (src->ships < fleet.ships) {
@@ -472,23 +472,23 @@ bool ParseBotFleets(GameState *state, Nob_String_View bot_message,
 
     nob_da_append(&state->fleets, fleet);
   }
-  nob_log(NOB_DEBUG, "done parsing bot %zu fleets", bot_idx);
+  nob_log(NOB_DEBUG, "done parsing bot %u fleets", bot_idx);
   return true;
 }
 
 void PrintBotDebugMessages(GameState *state, Nob_String_Builder *sb,
-                           size_t bot_idx) {
+                           unsigned bot_idx) {
   if (bot_idx >= state->bots.count) {
     nob_log(NOB_ERROR, "Tried accessing a bot OOB.");
     exit(1);
   }
   if (!TestBit(state->bot_bit_set, bot_idx) ||
       !subprocess_alive(state->bots.items[bot_idx].process)) {
-    nob_log(NOB_WARNING, "Bot %zu is not active.", bot_idx);
+    nob_log(NOB_WARNING, "Bot %u is not active.", bot_idx);
     return;
   }
 
-  const size_t max_chunk_length = 512;
+  const unsigned max_chunk_length = 512;
   sb->count = 0;
   bool message_ended = false;
 
@@ -509,16 +509,16 @@ void PrintBotDebugMessages(GameState *state, Nob_String_Builder *sb,
   }
 
   if (sb->count)
-    nob_log(NOB_INFO, "bot %zu says: |%.*s|", bot_idx, (unsigned)sb->count,
+    nob_log(NOB_INFO, "bot %u says: |%.*s|", bot_idx, (unsigned)sb->count,
             sb->items);
 }
 
 void RunBotCycle(GameState *state, Nob_String_Builder *bot_message) {
-  size_t bot_num = 0;
+  unsigned bot_num = 0;
   nob_da_foreach(Bot, bot, &state->bots) {
     // Skip disqualified or lost bots.
     if (TestBit(state->bot_bit_set, bot_num)) {
-      nob_log(NOB_DEBUG, "sending map to bot %zu", bot_num);
+      nob_log(NOB_DEBUG, "sending map to bot %u", bot_num);
 
       sendMapToBot(state, bot_num);
     }
@@ -542,7 +542,7 @@ void RunBotCycle(GameState *state, Nob_String_Builder *bot_message) {
       if (!bot_okay)
         DisqualifyBot(state, bot_num);
 
-      nob_log(NOB_DEBUG, "done with bot %zu, advancing to bot %zu", bot_num,
+      nob_log(NOB_DEBUG, "done with bot %u, advancing to bot %u", bot_num,
               bot_num + 1);
     }
     bot_num++;
@@ -665,7 +665,7 @@ int AdvanceTurn(GameState *state) {
   }
   // We do this after processing becuase we need to keep fleet order while
   // processing them.
-  for (size_t i = 0; i < state->fleets.count; i++) {
+  for (unsigned i = 0; i < state->fleets.count; i++) {
     if (state->fleets.items[i].remaining == 0) {
       nob_da_remove_unordered(&state->fleets, i);
       // Remove unordered replaces the current fleet with the last one,
@@ -719,7 +719,7 @@ void RunGame(GameState *state) {
   nob_sb_free(bot_message);
 }
 
-void UpdateStateFromLogEntry(GameState *state, size_t entry_idx) {
+void UpdateStateFromLogEntry(GameState *state, unsigned entry_idx) {
   if (entry_idx >= state->game_log.count) {
     nob_log(NOB_WARNING, "Attempted accessing OOB game log entry.");
     return;
